Refactored parts of InitializationData into MultiModel data class
Made InitializationData and MaestroV1SimulationConfiguration subclasses of MultiModel.
